+++ web/src/pages/index.tsx
@@ -1,0 +1,10 @@
+import { Container } from "../components/Container";
+import { DarkModeSwitch } from "../components/DarkModeSwitch";
+
+const Index = () => (
+  <Container height="100vh">
+    <DarkModeSwitch />
+  </Container>
+);
+
+export default Index;